requirements.txt
playwright==1.52.0
apscheduler==3.10.4
pyyaml==6.0.2
python-dotenv==1.0.1
